// Assets-Uncapped/front-portal-develop/src/services/api/partners/models/ApplicationDetailsResponse.ts
export interface ApplicationDetailsResponse {
    /**
     * 
     * @type {string}
     * @memberof ApplicationDetailsResponse
     */
    id?: string;
    /**
     * 
     * @type {Date}
     * @memberof ApplicationDetailsResponse
     */
    createDate?: Date;
    /**
     * 
     * @type {Date}
     * @memberof ApplicationDetailsResponse
     */
    updateDate?: Date;
    /**
     * 
     * @type {string}
     * @memberof ApplicationDetailsResponse
     */
    partnerId?: string;
    /**
     * 
     * @type {OfferDetailsResponse}
     * @memberof ApplicationDetailsResponse
     */
    offerDetailsResponse?: OfferDetailsResponse;
    /**
     * 
     * @type {ClientDetailsResponse}
     * @memberof ApplicationDetailsResponse
     */
    clientDetailsResponse?: ClientDetailsResponse;
    /**
     * 
     * @type {OrganisationDetailsResponse}
     * @memberof ApplicationDetailsResponse
     */
    organisationDetailsResponse?: OrganisationDetailsResponse;
    /**
     * 
     * @type {Array<DealDetailsResponse>}
     * @memberof ApplicationDetailsResponse
     */
    dealsDetailsResponse?: Array<DealDetailsResponse>;
    /**
     * 
     * @type {string}
     * @memberof ApplicationDetailsResponse
     */
    status?: ApplicationDetailsResponseStatusEnum;
    /**
     * 
     * @type {string}
     * @memberof ApplicationDetailsResponse
     */
    type?: ApplicationDetailsResponseTypeEnum;
    /**
     * 
     * @type {string}
     * @memberof ApplicationDetailsResponse
     */
    statusDetails?: string;
}


/**
 * @export
 */
export const ApplicationDetailsResponseStatusEnum = {
    Draft: 'DRAFT',
    NotEligible: 'NOT_ELIGIBLE',
    Submitted: 'SUBMITTED',
    Pending: 'PENDING',
    Duplicate: 'DUPLICATE',
    Invited: 'INVITED',
    Created: 'CREATED',
    Started: 'STARTED',
    Approved: 'APPROVED',
    Rejected: 'REJECTED',
    Active: 'ACTIVE',
    Closed: 'CLOSED',
    Abandoned: 'ABANDONED'
} as const;
export type ApplicationDetailsResponseStatusEnum = typeof ApplicationDetailsResponseStatusEnum[keyof typeof ApplicationDetailsResponseStatusEnum];

/**
 * @export
 */
export const ApplicationDetailsResponseTypeEnum = {
    Direct: 'DIRECT',
    Referral: 'REFERRAL',
    Import: 'IMPORT',
    Existing: 'EXISTING',
    External: 'EXTERNAL'
} as const;
export type ApplicationDetailsResponseTypeEnum = typeof ApplicationDetailsResponseTypeEnum[keyof typeof ApplicationDetailsResponseTypeEnum];